혁재 : 웹서버 수정, 불필요한 함수들 제거
